% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{ShinyMSMGOptimizer}
\alias{ShinyMSMGOptimizer}
\title{Launch the MSMG SimaPro CSV Optimizer}
\usage{
ShinyMSMGOptimizer(...)
}
\arguments{
\item{...}{Arguments passed to \code{shiny::runApp()}.}
}
\description{
Starts the MSMG SimaPro CSV Optimizer Shiny application.
}
